// frontend/src/api/model/domainResponse.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Whendoist
 * WHEN do I do my tasks?
 * OpenAPI spec version: 0.55.68
 */

/**
 * Response model for a domain.
 */
export interface DomainResponse {
  id: number;
  name: string;
  color: string | null;
  icon: string | null;
  position: number;
  is_archived: boolean;
}